object_detection_v3/requirements.txt
absl-py==0.6.1
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.65
botocore==1.12.65
brewer2mpl==1.4.1
bs4==0.0.1
bz2file==0.98
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
community==1.0.0b1
configparser==3.5.0
constantly==15.1.0
Crawl==0.5.5
cryptography==2.3
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
docutils==0.14
elasticsearch==6.3.0
elasticsearch-dsl==6.2.0
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Admin==1.5.2
flask-admin-elasticsearch-dsl==0.0.1
Flask-SQLAlchemy==2.3.2
gast==0.2.0
gensim==3.6.0
geographiclib==1.49
geopy==1.18.1
grpcio==1.16.0
h5py==2.8.0
happybase==1.1.0
hyperlink==18.0.0
idna==2.7
imageio==2.5.0
incremental==17.5.0
ipaddress==1.0.22
itsdangerous==0.24
jdcal==1.4
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lda==1.1.0
libsvm==3.22
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
networkx==2.2
nltk==3.4
numpy==1.15.2
opencv-python==4.1.0.25
openpyxl==2.5.14
pandas==0.23.4
parsel==1.5.0
pbr==3.1.1
Pillow==5.2.0
ply==3.11
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.2
pypiwin32==223
python-dateutil==2.7.3
python-modularity-maximization==0.0.1rc4
pytz==2018.5
PyWavelets==1.0.3
pywin32==224
PyYAML==3.13
queuelib==1.5.0
redis==2.10.6
requests==2.20.0
resolver==0.2.1
s3transfer==0.1.13
scikit-image==0.15.0
scikit-learn==0.20.0
scipy==1.1.0
Scrapy==1.5.1
scrapy-redis==0.6.8
ScrapyElasticSearch==0.9.1
seaborn==0.9.0
selector==0.10.1
service-identity==17.0.0
SimpleITK==1.2.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.7.1
soupsieve==1.9.1
SQLAlchemy==1.2.9
statsmodels==0.9.0
stop-words==2018.7.23
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
thriftpy==0.3.9
torch==1.0.1
torchvision==0.2.2.post3
tornado==5.1.1
Twisted==18.9.0
UNKNOWN==0.0.0
urllib3==1.24.1
w3lib==1.19.0
Werkzeug==0.14.1
wordcloud==1.5.0
WTForms==2.2.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
